Add model tests

+++ app/helpers/application_helper.rb
@@ -28,10 +28,13 @@ module ApplicationHelper
   end
   def show_vote_btn(article)
     vote = Vote.find_by(article: article, user: current_user)
-    if vote
-      link_to('Unvote', article_vote_path(id: vote.id, article_id: article.id), method: :delete, class: 'btn btn-danger')
+    if logged_in?
+      if vote
+        link_to('Unvote', article_vote_path(id: vote.id, article_id: article.id), method: :delete, class: 'btn btn-danger')
+      else
+        link_to('Vote', article_votes_path(article_id: article.id), method: :post, class: 'btn btn-primary')
+      end
     else
-      link_to('Vote', article_votes_path(article_id: article.id), method: :post, class: 'btn btn-primary')
     end
   end
 end
